backend/requirements.txt
ably==2.1.1
anyio==4.11.0
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
filelock==3.19.1
fsspec==2025.9.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.0
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.2
methoddispatch==3.0.2
mpmath==1.3.0
msgpack==1.1.2
networkx==3.5
nltk==3.9.1
numpy==2.3.3
packaging==25.0
pillow==11.3.0
pyee==12.1.1
PyPDF2==3.0.1
python-multipart==0.0.20
PyYAML==6.0.2
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.0
setuptools==80.9.0
sniffio==1.3.1
sqlparse==0.5.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.7.1+cpu
tqdm==4.67.1
transformers==4.56.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
